+++ scripts/public-additional-scripts.js
@@ -6,6 +6,7 @@ import './public/landing-banners';
 import './public/navigation';
 import './public/search';
 import './public/side-scroll-generator';
+import './public/tables';
 
 import './public/cleanbc/accessibility';
 import './public/betterhomes/accessibility';